+++ cypress.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'cypress';
 import { config } from './package.json';
 
 export default defineConfig({
-  retries: 4,
+//  retries: 4,
   videosFolder: 'test/cypress/videos',
   video: false,
   screenshotsFolder: 'test/cypress/screenshots',